+++ igrep/tests/test_index.cpp
@@ -5,6 +5,7 @@
 #include<fstream>
 #include<string>
 #include<vector>
+#include<cstdint>
 
 using namespace std;
 using namespace std::filesystem;
@@ -12,7 +13,7 @@ using namespace igrep::indexer;
 
 TEST(IndexTest, ProcessLine_GetPositions_Success){
     Index index;
-    size_t word_index = 1;
+    uint32_t word_index = 1;
     string line = "Hello, world";
     path filepath = "/path/to/file.txt";
 
@@ -36,8 +37,8 @@ TEST(IndexTest, GetPositions_NonExistentWord_ReturnEmpty){
 TEST(IndexTest, SerializeIndex_ThenDeserializeToAnother_SameIndexes){
     Index index;
     string line = "Hello, world";
-    size_t first_file_index = 1;
-    size_t second_file_index = 1;
+    uint32_t first_file_index = 1;
+    uint32_t second_file_index = 1;
     index.process_line(line, "/path/to/file1.txt", 1, first_file_index);
     index.process_line(line, "/path/to/file2.txt", 1, second_file_index);
     path index_path = string(SOURCE_DIR) + "/testdata/index.bin";
@@ -67,9 +68,9 @@ TEST(IndexTest, DeleteFile_IndexedFile_CorrectIndex){
     string second_file = "/path/to/file2.txt";
     string first_string = "First file string";
     string second_string = "Second file string";
-    size_t first_file_index = 1;
-    size_t second_file_index = 1;
-	size_t expected_word_index = 1;
+    uint32_t first_file_index = 1;
+    uint32_t second_file_index = 1;
+	uint32_t expected_word_index = 1;
     index.process_line(first_string, first_file, 1, first_file_index);
     index.process_line(second_string, second_file, 1, second_file_index);
     expected_index.process_line(second_string, second_file, 1, expected_word_index);
